src/VumbaSoft.AdventureWorks.Controllers/Sales/Individuals/IndividualsController.cs
using Microsoft.AspNetCore.Mvc;
using VumbaSoft.AdventureWorks.Objects;
using VumbaSoft.AdventureWorks.Services;
using VumbaSoft.AdventureWorks.Validators;
using System;

namespace VumbaSoft.AdventureWorks.Controllers.Sales
{
    [Area("Sales")]
    public class IndividualsController : ValidatedController<IIndividualValidator, IIndividualService>
    {
        public IndividualsController(IIndividualValidator validator, IIndividualService service)
            : base(validator, service)
        {
        }

        [HttpGet]
        public ViewResult Index()
        {
            return View(Service.GetViews());
        }

        [HttpGet]
        public ViewResult Create()
        {
            return View();
        }

        [HttpPost]
        public ActionResult Create(IndividualView individual)
        {
            if (!Validator.CanCreate(individual))
                return View(individual);

            Service.Create(individual);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Details(Int32 id)
        {
            return NotEmptyView(Service.Get<IndividualView>(id));
        }

        [HttpGet]
        public ActionResult Edit(Int32 id)
        {
            return NotEmptyView(Service.Get<IndividualView>(id));
        }

        [HttpPost]
        public ActionResult Edit(IndividualView individual)
        {
            if (!Validator.CanEdit(individual))
                return View(individual);

            Service.Edit(individual);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Delete(Int32 id)
        {
            return NotEmptyView(Service.Get<IndividualView>(id));
        }

        [HttpPost]
        [ActionName("Delete")]
        public RedirectToActionResult DeleteConfirmed(Int32 id)
        {
            Service.Delete(id);

            return RedirectToAction("Index");
        }
    }
}
